+++ include/pajlada/settings/setting.hpp
@@ -288,6 +288,10 @@ public:
     bool
     setValue(const Type &newValue, SignalArgs &&args = SignalArgs())
     {
+        if (this->optionEnabled(SettingOption::CompareBeforeSet)) {
+            args.compareBeforeSet = true;
+        }
+
         {
             std::unique_lock<std::mutex> lock(this->valueMutex);
             this->value = newValue;